Added envify for easy production / development configuration
