+++ server/package.json
@@ -23,6 +23,7 @@
   "devDependencies": {
     "browserify": "^8.1.3",
     "connect-livereload": "^0.5.3",
+    "envify": "^3.2.0",
     "tiny-lr": "^0.1.5",
     "watchify": "^2.3.0"
   },
